images upload and edit

+++ utils/helpers.js
@@ -5,20 +5,20 @@ module.exports = {
 
     // determines is the student is ready to graduate
     grad: (a,b,c,d,e,f,g,h,i,j,k,l,m,n) => {
-      a = parseInt(a);
-      b = parseInt(b);
-      c = parseInt(c);
-      d = parseInt(d);
-      e = parseInt(e);
-      f = parseInt(f);
-      g = parseInt(g);
-      h = parseInt(h);
-      i = parseInt(i);
-      j = parseInt(j);
-      k = parseInt(k);
-      l = parseInt(l);
-      m = parseInt(m);
-      n = parseInt(n);
+      a = parseFloat(a);
+      b = parseFloat(b);
+      c = parseFloat(c);
+      d = parseFloat(d);
+      e = parseFloat(e);
+      f = parseFloat(f);
+      g = parseFloat(g);
+      h = parseFloat(h);
+      i = parseFloat(i);
+      j = parseFloat(j);
+      k = parseFloat(k);
+      l = parseFloat(l);
+      m = parseFloat(m);
+      n = parseFloat(n);
       
       if (a >= b && c >= d && e >= f && g >= h && i >= j && k >= l && m >= n) {
         return ('<span style="color: green">Yes!</span>');
@@ -30,8 +30,8 @@ module.exports = {
 
 // greater than or equal to
     ge: (a,b) => {
-      a = parseInt(a);
-      b = parseInt(b);
+      a = parseFloat(a);
+      b = parseFloat(b);
       
       if (a < b) {
         return ('<span style="color: red">');
